LPS-47201 These test log classes are not utilities that should be reused outside portal, so it doesn't make sense having them under portal-test. Moving them to portal-impl/test/integration
