+++ portal-impl/test/integration/com/liferay/portal/log/Log4JLoggerTestUtil.java
@@ -12,12 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.log.test;
+package com.liferay.portal.log;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.log.LogWrapper;
-import com.liferay.portal.log.Log4jLogImpl;
 
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
