+++ portal-impl/test/integration/com/liferay/portal/spring/transaction/TransactionInterceptorTest.java
@@ -19,8 +19,8 @@ import com.liferay.portal.kernel.bean.PortalBeanLocatorUtil;
 import com.liferay.portal.kernel.dao.orm.EntityCacheUtil;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.util.InfrastructureUtil;
-import com.liferay.portal.log.test.CaptureAppender;
-import com.liferay.portal.log.test.Log4JLoggerTestUtil;
+import com.liferay.portal.log.CaptureAppender;
+import com.liferay.portal.log.Log4JLoggerTestUtil;
 import com.liferay.portal.model.ClassName;
 import com.liferay.portal.model.impl.ClassNameImpl;
 import com.liferay.portal.service.ClassNameLocalServiceUtil;
